+++ cli/src/templates/router/domain/types/{{camel_case_name}}EventRecord.types.ts
@@ -1,6 +1,6 @@
 {{#is_worker}}import type { WorkerEventEntity } from '@forklaunch/interfaces-worker/types';
 
-export interface I{{pascal_case_name}}EventRecord extends WorkerEventEntity {
+export interface {{pascal_case_name}}EventRecord extends WorkerEventEntity {
   message: string;
 }
 {{/is_worker}}