+++ blueprint/implementations/billing/base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/implementation-billing-base
 
+## 1.0.4
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/interfaces-billing@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes