+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/hyper-express
 
+## 1.1.3
+
+### Patch Changes
+
+- Package versions and simplified compliance entity typing
+- Updated dependencies
+  - @forklaunch/validator@1.1.3
+  - @forklaunch/common@1.1.3
+  - @forklaunch/core@1.1.3
+  - @forklaunch/ws@1.1.3
+
 ## 1.1.2
 
 ### Patch Changes