+++ framework/validator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/validator
 
+## 1.1.3
+
+### Patch Changes
+
+- Package versions and simplified compliance entity typing
+- Updated dependencies
+  - @forklaunch/common@1.1.3
+
 ## 1.1.2
 
 ### Patch Changes